Added copy_for_reshape to contain implementation
